requirements.txt
httpx
selectolax
selenium
python-dotenv
pandas